requirements.txt
tensorflow-gpu==1.12.0
numpy
pandas
scikit-learn
tabulate
tqdm
comet_ml
joblib
